Add SetAirplayVolume to the D-Bus RemoteControl interface. Add SetVolume to the MPRIS interface. Hook up the Volume property in the MPRIS interface. Modify RemoteCommand in the D-Bus interface to return the HTTP status and response. Change the type of airplay_volume from int to double in the metadata hub. Add a few sample commands in the D-Bus document.

+++ metadata_hub.h
@@ -129,7 +129,7 @@ typedef struct metadata_bundle {
 
   int speaker_volume; // this is the actual speaker volume, allowing for the main volume and the
                       // speaker volume control
-  int airplay_volume;
+  double airplay_volume;
 
   metadata_watcher watchers[number_of_watchers]; // functions to call if the metadata is changed.
   void *watchers_data[number_of_watchers];       // their individual data
